Export nifti related function in a separate file
